Remove unused specifications, consolidate the remaining ones for use by porter alone.
