+++ tests/unit/test_character_serialization.py
